# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for the academic paper formatting assistant
servers:
  - url: /api
    description: Base API path
...
security:
  - cookieAuth: []
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Create an account and start a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created; a session cookie is set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invalid registration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Sign in and start a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in; a session cookie is set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Email or password incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the signed-in user
      responses:
        "200":
          description: The current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the signed-in user's password
      description: >-
        Revokes every other session for the user and re-issues one for the
        calling device.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "204":
          description: Password changed
        "401":
          description: Current password incorrect, or not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: New password does not meet requirements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pf_session
      description: >-
        httpOnly session cookie issued by /auth/login and /auth/register.
        Browsers attach it automatically; requests must be sent with
        credentials included.
  schemas:
...
    User:
      type: object
      required: [id, email, displayName, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        displayName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 8
          maxLength: 200
        displayName:
          type: string
          minLength: 1
          maxLength: 120
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          maxLength: 200
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
          maxLength: 200
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...